requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
redis==5.0.1
aioredis==2.0.1
pydantic[email]
plaid-python
alembic
email-validator
bcrypt
pytest
fastapi-users[sqlalchemy]
authlib
itsdangerous 
pyjwt 